Initial POC for slim-lint
This is a working proof-of-concept for linting Slim templates with
RuboCop. It borrows heavily from haml-lint project for most of the
configuration system and other plumbing boilerplate.

The system for specifying linters tries to provide a nice easy-to-read
DSL, but it's not clear to me if this is the best way to go about it
(especially since you cannot pass blocks to other blocks).

Unfortunately, the underlying Temple framework that both Slim and thus
this project are now dependent on does not retain detailed source map
information, so the best granularity we will get is at the line level
(and even then it may be off by one or two lines depending on how the
syntax tree is generated, but those edge cases remain to be discovered).