Wrap Sexp atoms in Atom class
One of the issues we're already seeing is the need to explicitly check
whether a Sexp is an expression or an atom. Rather than have to do this
check every time, we create an `Atom` class which wraps the literal
value and defines `==` and `match?` methods so it acts like a Sexp where
it matters.

This allows us to simplify logic to not concern ourselves with whether
an object is an expression or an atom--the underlying helpers will do
the differentiation for us.

As part of this change, a number of other areas were changed so things
worked--delegating methods to the underlying value requires a slight
change in strategy when doing regex matches, for example.

We also add some tests for the `Sexp` class.