Removed unused configuration functionality
We don't use the `ALL` hook anywhere, and there's not much practical
purpose to it. We can easily add it back if a need arises again.

Also, don't auto-convert `nil`s in the configuration to empty hashes, as
this also isn't used and it seems like it could be a source of confusion
for users who explicitly set a configuration value to `nil`.