Allow linters to include/exclude files in config
Add support for `include`/`exclude` options which allow a user to filter
or reject files from the scope of a specific linter.

In the process of implementing this, it was realized that the `Runner`
class was taking on too much responsibility, and was also difficult to
test. This prompted the extraction of the `LinterSelector` class, which
now handles the task of filtering linters based on given command line
options and configuration for those linters.