Improve Ruby code extraction
We previously made the assumption that utilizing the built-in Slim
filters would be sufficient for extracting Ruby code. Unfortunately, it
turns out that these filters inject a number of special Slim-specific
code in order to make the templates actually executable.

Since we're not interested in compiling a template that is actually
executable, swap out some Slim filters with our own modified versions
that perform less magic.

The end result is Ruby code that produces fewer erroneous RuboCop
warnings.