Gracefully handle encoding mishaps
We are getting reports like in #14 where we're not selecting the best
encoding for a file in some cases. While this is usually user error, we
can try to fix things by attempting to change the encoding to UTF-8 as
that's usually what is intended.

Piggy-back on the Temple library's `Encoding` filter to clean this up
for us.