Fix exclude option to work with "./" in paths
When specifying a `.` as an argument to the list of files slim-lint
should search, the paths returned would include that `.` so that paths
would look like `./something.slim` instead of just `something.slim`.

This broke exclusion since developers would write the path without the
leading `./`.

Fix this by normalizing the paths to not have the leading `./`.