Update minimum RuboCop version to 0.36.0
RuboCop's recent removal of the Astrolabe gem dependency[1] meant that
we could no longer require it ourselves. It appears that the Astrolabe
gem isn't accepting pull requests, so the RuboCop maintainers chose to
inline the relevant code directly into RuboCop.

Since slim-lint is tightly-coupled to RuboCop, we're better-served by
using RuboCop's implementation than Astrolabes, so switch to it.

In the process, we had to fix a number of new warnings reported by cops
introduced in the new version.