Overhaul Ruby Extraction
This change introduces a new parser, based on the Slim parser from v4.0.1, that deliberately tracks line numbers, column numbers, and indentation (for most multiline Ruby code) as it parses. Previously, we'd tried to work backwards from the parse tree to infer line numbers, but this approach was inherently limited and could not be coerced to handle simple cases (e.g. an embedded `ruby:` block followed by an arbitrary number of newlines).

By tracking source location information from the start, we get better quality data, and can produce more representative Ruby extractions. So much better, in fact, that we no longer need to ignore Rubocop violations in order to avoid serialization errors.

Unfortunately, since the entire Temple (and therefore Slim) toolchain expects _Arrays_ rather than our internal Sexp class, most of the Slim parsing pipeline needed to be re-implemented to work properly with the new data structures. At this point, we have verification that the test cases in the Slim literate test file all parse identically (modulo where we deliberately diverge re: whitespace), and the existing test suite also passes. This should not be taken as an indication that the line and column numbering is _correct_, simply that we haven't broken existing test code.