Add Strict locals missing linter (#175)
And here I propose another linter that we would love to use in our
project. It checks for [Strict
locals](https://guides.rubyonrails.org/action_view_overview.html#strict-locals)
magic line and reports when the magic line is missing from the template.
Strict locals are encouraged by some in Rails partials as they allow to
define an explicit "API" for the templates showing clearly which
variables should be passed into the template when rendering it vs. which
ones are purely local for the template. They also allow to define
default values. I wrote a little
[post](https://dev.to/nejremeslnici/strict-locals-in-slim-haml-partials-in-rails-2f73)
about them recently. Thanks for your considering this!


![image](https://github.com/sds/slim-lint/assets/462701/c34333ec-4dbe-42f4-8e34-fabf879ecc69)


![image](https://github.com/sds/slim-lint/assets/462701/96c491ee-4ccd-44fe-8bcb-4227770af5c5)